\name{ALLModules}
\docType{data}
\alias{ALLModules}
\alias{ALLModulesSmall}
\title{ISA transcription modules for the ALL data}
\description{
  The Iterative Signature Algorithm (ISA) is a biclustering
  method. \code{ALLModules} and \code{ALLModulesSmall} are example ISA
  biclusters (=modules) found in the ALL data set.
}
\usage{
ALLModules
ALLModulesSmall
}
\format{Both \code{ALLModules} and \code{ALLModulesSmall} are instances
  of the \code{ISAModules} class.}
\source{\code{ISAModules} was generated by calling \code{ISA} on the
  \code{ALL} data set, using the default
  parameters. \code{ISAModulesSmall} was generated the same way, but
  with gene threshold 2.7 and condition threshold 1.4 only.}
\references{
  Bergmann S, Ihmels J, Barkai N: Iterative signature algorithm for the
  analysis of large-scale gene expression data \emph{Phys Rev E Stat
    Nonlin Soft Matter Phys.} 2003 Mar;67(3 Pt 1):031902. Epub 2003 Mar 11.
}
\seealso{The \code{ALL} BioConductor package.}
\examples{
data(ALLModules)
ALLModules
}
\keyword{datasets}
