\name{ISA.html}
\alias{ISA.html}
\alias{ISA.html.modules}
\alias{ISA.html.table}
\title{Create HTML summary pages from the result of modular analysis}
\description{
  These functions create various sophisticated HTML pages from a set of
  ISA biclusters.
}
\usage{
ISA.html.table (modules, target.dir, which = seq_len(length(modules)), 
    template = system.file("autogen", package = "eisa"), GO = NULL, 
    KEGG = NULL, miRNA = NULL, CHR = NULL, htmltitle = NULL, 
    notes = NULL, seed = NULL) 

ISA.html.modules1 (eset, modules, which = seq_len(length(modules)), target.dir, 
    template = system.file("autogen", package = "eisa"), GO = NULL,    
    KEGG = NULL, miRNA = NULL, CHR = NULL, cond.to.include = NULL,     
    cond.col = "white", sep = NULL, seed = NULL)                       

ISA.html (eset, modules, target.dir, template = system.file("autogen", 
    package = "eisa"), GO, KEGG, miRNA = NULL, CHR = NULL, htmltitle = NULL, 
    notes = NULL, seed = NULL, cond.to.include = NULL, cond.col = "white", 
    sep = NULL) 
}
\arguments{
  \item{modules}{An \code{ISAModules} object.}
  \item{target.dir}{Character vector of length one, the directory in
    which the result is placed. It is created if it does not exist.}
  \item{which}{Numeric vector, which modules to include in the table
    (for \code{ISA.html.table}); or, which modules to
    create HTML pages for (\code{ISA.html.modules}). All modules are
    used by default.}
  \item{template}{The directory containing the HTML template files. By
    default the template included in the \code{eisa} package is used.}
  \item{GO}{List of three \code{\link{GOListHyperGResult}} objects, as
    returned by the \code{\link{ISA.GO}} function.}
  \item{KEGG}{A \code{\link{KEGGListHyperGResult}} object, usually the
    output of the \code{\link{ISA.KEGG}} function.}
  \item{miRNA}{A \code{\link{miRNAListHyperGResult}} object, or
    \code{NULL}. See also the \code{\link{isa.miRNA}} function.}
  \item{CHR}{A \code{\link{CHRListHyperGResult}} object or \code{NULL},
    see also the \code{\link{ISA.CHR}} function. }
  \item{htmltitle}{Character vector of length one, the title of the HTML
    page.}
  \item{notes}{Character vector of length one. Optional HTML text, on
    the deafult template it is placed on the top of the page, above the
    table.}
  \item{seed}{Either \code{NULL}, or a character vector, with optional
    column that is added to the module table.}
  \item{eset}{An \code{ExpressionSet} or \code{ISAExpressionSet}
    object. If an \code{ExpressionSet} object is supplied, then it is
    normalised by calling \code{\link{ISA.normalize}} on it.}
  \item{cond.to.include}{Numeric or character vector, specifies which
    columns of the phyenotype data of the original expression matrix are
    included in the tables of samples. By default the first six columns
    are included.}
  \item{cond.col}{This is passed to \code{\link{cond.plot}} as the
    \code{col} argument.}
  \item{sep}{This is passed to \code{\link{cond.plot}} as the \code{sep}
    argument.}
}
\details{
  \code{ISA.html.table} creates an HTML page, a summary of the results
  of the modular analysis, including enrichment analysis of the
  modules.
  
  \code{ISA.html.modules} creates a separate HTML page for each module,
  including the following elements:
  \itemize{
    \item An expression plot of the genes and samples in the module,
      including the ISA scores. This is done by calling
      \code{\link{expPlot}}.
    \item Overlap plot of all modules, see \code{\link{overlap.plot}}
    \item Gene Ontology tree plots for the enriched GO terms, separately
      for the three ontologies., These are produced by calling
      \code{\link{goplot}}.
    \item Tables for the enriched Gene Ontology terms, separately for
      the three ontologies.
    \item A table for the enriched KEGG pathways.
    \item A table for the enriched miRNA families.
    \item The list of genes in the module.
    \item The list of samples in the module.
    \item A condition plot, see \code{\link{cond.plot}}.
  }

  By default, clicking on the lines of the table generated by
  \code{ISA.html.table} is linked to the HTML page of the module,
  generated by \code{ISA.html.modules}.

  \code{ISA.html} calls both \code{ISA.html.table} and
  \code{ISA.html.modules}.
}
\value{
  These functions do not return a value. (They return \code{NULL}, invisibly.)
}
\author{ Gabor Csardi \email{Gabor.Csardi@unil.ch} }
\references{
  Bergmann S, Ihmels J, Barkai N: Iterative signature algorithm for the
  analysis of large-scale gene expression data \emph{Phys Rev E Stat
    Nonlin Soft Matter Phys.} 2003 Mar;67(3 Pt 1):031902. Epub 2003 Mar 11.
}
\seealso{
  The vignette included in the \code{eisa} package.
}
\examples{}
\keyword{cluster}
