\name{ISA.KEGG}
\alias{ISA.KEGG}
\concept{Enrichment analysis}
\title{Calculate KEGG Pathway enrichment for transcription modules}
\description{
  KEGG pathway enrichment is calculated for each ISA module
  separately. In the end the result is corrected for multiple hypothesis
  testing.
}
\usage{}
\arguments{}
\details{}
\value{}
\author{}
\references{}
\seealso{}
\examples{}
\keyword{cluster}
