\name{ListHyperGParams-class}
\docType{class}
\alias{CHRListHyperGParams-class}
\alias{GOListHyperGParams-class}
\alias{KEGGListHyperGParams-class}
\alias{miRNAListHyperGParams-class}
\alias{drive}
\alias{drive<-}
\alias{drive,GOListHyperGParams,logical-method}
\alias{drive,KEGGListHyperGParams,logical-method}
\alias{drive,CHRListHyperGParams,logical-method}
\alias{drive,miRNAListHyperGParams,logical-method}
\alias{drive,GOListHyperGParams-method}
\alias{drive,KEGGListHyperGParams-method}
\alias{drive,CHRListHyperGParams-method}
\alias{drive,miRNAListHyperGParams-method}
\alias{drive<-,GOListHyperGParams,logical-method}
\alias{drive<-,KEGGListHyperGParams,logical-method}
\alias{drive<-,CHRListHyperGParams,logical-method}
\alias{drive<-,miRNAListHyperGParams,logical-method}
\alias{condGeneIdUniverse,CHRListHyperGResult-method}
\alias{condGeneIdUniverse,KEGGListHyperGResult-method}
\alias{condGeneIdUniverse,GOListHyperGResult-method}
\alias{condGeneIdUniverse,miRNAListHyperGResult-method}
\alias{conditional,GOListHyperGResult-method}
\alias{conditional,GOListHyperGParams-method}
\alias{conditional<-,GOListHyperGParams,logical-method}

\alias{expectedCounts,CHRListHyperGResult-method}
\alias{expectedCounts,KEGGListHyperGResult-method}
\alias{expectedCounts,GOListHyperGResult-method}
\alias{expectedCounts,miRNAListHyperGResult-method}

\alias{geneCounts,CHRListHyperGResult-method}
\alias{geneCounts,KEGGListHyperGResult-method}
\alias{geneCounts,GOListHyperGResult-method}
\alias{geneCounts,miRNAListHyperGResult-method}

\alias{geneIdsByCategory,CHRListHyperGResult-method}
\alias{geneIdsByCategory,KEGGListHyperGResult-method}
\alias{geneIdsByCategory,GOListHyperGResult-method}
\alias{geneIdsByCategory,miRNAListHyperGResult-method}

\alias{geneIdUniverse,CHRListHyperGResult-method}
\alias{geneIdUniverse,KEGGListHyperGResult-method}
\alias{geneIdUniverse,GOListHyperGResult-method}
\alias{geneIdUniverse,miRNAListHyperGResult-method}

\alias{geneMappedCount,CHRListHyperGResult-method}
\alias{geneMappedCount,KEGGListHyperGResult-method}
\alias{geneMappedCount,GOListHyperGResult-method}
\alias{geneMappedCount,miRNAListHyperGResult-method}

\alias{htmlReport,CHRListHyperGResult-method}
\alias{htmlReport,KEGGListHyperGResult-method}
\alias{htmlReport,GOListHyperGResult-method}
\alias{htmlReport,miRNAListHyperGResult-method}

\alias{hyperGTest,CHRListHyperGParams-method}
\alias{hyperGTest,KEGGListHyperGParams-method}
\alias{hyperGTest,GOListHyperGParams-method}
\alias{hyperGTest,miRNAListHyperGParams-method}

\alias{makeValidParams,CHRListHyperGParams-method}
\alias{makeValidParams,KEGGListHyperGParams-method}
\alias{makeValidParams,GOListHyperGParams-method}
\alias{makeValidParams,miRNAListHyperGParams-method}

\alias{oddsRatios,CHRListHyperGResult-method}
\alias{oddsRatios,KEGGListHyperGResult-method}
\alias{oddsRatios,GOListHyperGResult-method}
\alias{oddsRatios,miRNAListHyperGResult-method}

\alias{ontology,GOListHyperGResult-method}
\alias{ontology<-,GOListHyperGParams,character-method}
\alias{ontology,GOListHyperGParams-method}

\alias{pvalues,CHRListHyperGResult-method}
\alias{pvalues,KEGGListHyperGResult-method}
\alias{pvalues,GOListHyperGResult-method}
\alias{pvalues,miRNAListHyperGResult-method}

\alias{sigCategories,CHRListHyperGResult-method}
\alias{sigCategories,KEGGListHyperGResult-method}
\alias{sigCategories,GOListHyperGResult-method}
\alias{sigCategories,miRNAListHyperGResult-method}

\alias{summary,CHRListHyperGResult-method}
\alias{summary,KEGGListHyperGResult-method}
\alias{summary,GOListHyperGResult-method}
\alias{summary,miRNAListHyperGResult-method}

\alias{universeCounts,CHRListHyperGResult-method}
\alias{universeCounts,KEGGListHyperGResult-method}
\alias{universeCounts,GOListHyperGResult-method}
\alias{universeCounts,miRNAListHyperGResult-method}

\alias{universeMappedCount,CHRListHyperGResult-method}
\alias{universeMappedCount,KEGGListHyperGResult-method}
\alias{universeMappedCount,GOListHyperGResult-method}
\alias{universeMappedCount,miRNAListHyperGResult-method}

\title{Classes for GO/KEGG/CHR/miRNA enrichment of ISA modules}
\description{}
\section{Methods}{
}
\seealso{}
\examples{}
\keyword{cluster}
