\name{mnplot}
\alias{mnplot}
\alias{ISAmnplot}
\title{Plot group means against each other for an ISA module}
\description{Plot mean expression values for two sets of samples,
  against each other.}
\usage{
mnplot (x, eset, data, group, \dots)
ISAmnplot (modules, number, eset, norm = c("raw", "feature", "sample"),
     data = annotation(modules), group, \dots)
}
\arguments{
  \item{x}{}
  \item{eset}{}
  \item{data}{}
  \item{group}{}
  \item{\dots}{}
  \item{modules}{}
  \item{number}{}
  \item{norm}{}
}
\details{}
\value{}
\author{}
\references{}
\seealso{The \code{\link[annotate]{GOmnplot}} and
  \code{\link[annotate]{KEGGmnplot}} functions in the \code{annotate}
  package.}
\examples{}
\keyword{cluster}
