\name{ISAHTML}
\alias{ISAHTML}
\alias{ISAHTMLModules}
\alias{ISAHTMLTable}
\title{Create HTML summary pages from the result of modular analysis}
\description{
  These functions create various sophisticated HTML pages from a set of
  ISA biclusters.
}
\usage{
ISAHTMLTable (modules, target.dir, which = NULL,
    template = system.file("autogen", package = "eisa"), GO = NULL, 
    KEGG = NULL, miRNA = NULL, CHR = NULL, htmltitle = NULL, 
    notes = NULL, seed = NULL, extra = list()) 

ISAHTMLModules (eset, modules, which = NULL, target.dir, 
    template = system.file("autogen", package = "eisa"), GO = NULL,    
    KEGG = NULL, miRNA = NULL, CHR = NULL, cond.to.include = NULL,     
    cond.col = "white", sep = NULL, seed = NULL, condPlot = TRUE)

ISAHTML (eset, modules, target.dir, template = system.file("autogen", 
    package = "eisa"), GO, KEGG, miRNA = NULL, CHR = NULL, htmltitle = NULL, 
    notes = NULL, seed = NULL, table.extra = list(), cond.to.include = NULL, 
    cond.col = "white", sep = NULL, condPlot = TRUE, which = NULL)
}
\arguments{
  \item{modules}{An \code{ISAModules} object.}
  \item{target.dir}{Character vector of length one, the directory in
    which the result is placed. It is created if it does not exist.}
  \item{which}{Numeric vector, which modules to include in the table
    (\code{ISAHTMLTable}); or, which modules to
    create HTML pages for (\code{ISAHTML} and \code{ISAHTMLModules}).
    All modules are used if this argument is \code{NULL}, 
    which is the default.}
  \item{template}{The directory containing the HTML template files. By
    default the template included in the \code{eisa} package is used.}
  \item{GO}{List of three \code{\link{GOListHyperGResult}} objects, as
    returned by the \code{\link{ISAGO}} function.}
  \item{KEGG}{A \code{\link{KEGGListHyperGResult}} object, usually the
    output of the \code{\link{ISAKEGG}} function.}
  \item{miRNA}{A \code{\link{miRNAListHyperGResult}} object, or
    \code{NULL}. See also the \code{\link{ISAmiRNA}} function.}
  \item{CHR}{A \code{\link{CHRListHyperGResult}} object or \code{NULL},
    see also the \code{\link{ISACHR}} function. }
  \item{htmltitle}{Character vector of length one, the title of the HTML
    page.}
  \item{notes}{Character vector of length one. Optional HTML text, on
    the default template it is placed on the top of the page, above the
    table.}
  \item{seed}{Either \code{NULL}, or a character vector, with an optional
    column that is added to the module table.}
  \item{extra}{Extra columns to put in the HTML table. It should be a
    named list of character vectors, each with the same length as the
    number of modules.}
  \item{table.extra}{This is passed to \code{ISAHTMLTable} as the
    \code{extra} argument.}
  \item{eset}{An \code{ExpressionSet} or \code{ISAExpressionSet}
    object. If an \code{ExpressionSet} object is supplied, then it is
    normalised by calling \code{\link{ISANormalize}} on it.}
  \item{cond.to.include}{Numeric or character vector, specifies which
    columns of the phyenotype data of the original expression matrix are
    included in the tables of samples. By default the first six columns
    are included.}
  \item{cond.col}{This is passed to \code{\link{condPlot}} as the
    \code{col} argument.}
  \item{sep}{This is passed to \code{\link{condPlot}} as the \code{sep}
    argument.}
  \item{condPlot}{Logical scalar, whether to create condition plots.
    If an alternative biclustering method was used to find the modules,
    then probably it makes no sense creating condition plots for them.}
}
\details{
  \code{ISAHTMLTable} creates an HTML page, a summary of the results
  of the modular analysis, including enrichment analysis of the
  modules.
  
  \code{ISAHTMLModules} creates a separate HTML page for each module,
  including the following elements:
  \itemize{
    \item An expression plot of the genes and samples in the module,
      including the ISA scores. This is done by calling
      \code{\link{expPlot}}.
    \item Gene Ontology tree plots for the enriched GO terms, separately
      for the three ontologies. These are produced by calling
      \code{\link{gograph}}.
    \item Tables for the enriched Gene Ontology terms, separately for
      the three ontologies.
    \item A table for the enriched KEGG pathways.
    \item A table for the enriched miRNA families.
    \item The list of genes in the module.
    \item The list of samples in the module.
    \item A condition plot (if the \code{condPlot} argument is
    \code{TRUE}), see \code{\link{condPlot}}. 
  }

  By default, clicking on the rows of the table generated by
  \code{ISAHTMLTable} is linked to the HTML page of the module,
  generated by \code{ISAHTMLModules}.

  \code{ISAHTML} calls both \code{ISAHTMLTable} and
  \code{ISAHTMLModules}.
}
\value{
  These functions do not return a value. (They return \code{NULL}, invisibly.)
}
\author{ Gabor Csardi \email{csardi.gabor@gmail.com} }
\references{
  Bergmann S, Ihmels J, Barkai N: Iterative signature algorithm for the
  analysis of large-scale gene expression data \emph{Phys Rev E Stat
    Nonlin Soft Matter Phys.} 2003 Mar;67(3 Pt 1):031902. Epub 2003 Mar 11.
}
\seealso{
  The vignette included in the \code{eisa} package.
}
\examples{
# Load data
library(ALL)
data(ALL)
data(ALLModulesSmall)

# Calculate enrichment
GO <- ISAGO(ALLModulesSmall)
KEGG <- ISAKEGG(ALLModulesSmall)
CHR <- ISACHR(ALLModulesSmall)

# Generate HTML summary
htmldir <- tempdir()
ISAHTML(ALL, modules=ALLModulesSmall, target.dir=htmldir,
        GO=GO, KEGG=KEGG, CHR=CHR)

# Open a browser to view the summary
if (interactive()) {
  browseURL(URLencode(paste("file://", htmldir, "/maintable.html", sep="")))
}
}
\keyword{cluster}
