\name{ISAExpressionSet-class}
\docType{class}
\alias{ISAExpressionSet-class}
\alias{feat.exprs,ISAExpressionSet}
\alias{samp.exprs,ISAExpressionSet}
\alias{hasNA,ISAExpressionSet}
\alias{hasNA<-,ISAExpressionSet}
\alias{prenormalized,ISAExpressionSet}
\alias{prenormalized<-,ISAExpressionSet}
\title{Expression Set, normalized for using with ISA}
\description{}
\section{Methods}{
}
\seealso{}
\examples{}
\keyword{classes}
