\name{ListHyperGParams-class.Rd}
\docType{class}
\alias{CHRListHyperGParams-class}
\alias{DBDListHyperGParams-class}
\alias{GOListHyperGParams-class}
\alias{KEGGListHyperGParams-class}
\alias{miRNAListHyperGParams-class}
\title{Classes for GO/KEGG/CHR/miRNA enrichment of ISA modules}
\description{}
\section{Methods}{
}
\seealso{}
\examples{}
\keyword{cluster}
